pyproject.toml
[tool.poetry]
name = "meimei"
version = "0.0.0"
description = "A web app interfacing an AI that can spot cats."
authors = ["petejh <petejh.code@q.com>"]
license = "MIT"

[tool.poetry.dependencies]
# upper bound on Python version is due to numpy
python = "^3.9, <3.11"
flask = "^1.1.2"
deepen = {git = "https://github.com/petejh/deepen.git"}
numpy = "^1.21.2"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^0.17"
h5py = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
